NEW UPDATE; NeoCity, 2142
Wow, I'm off school today because of polling in the UK so I'm back at it again. Coding this is draining yet rewarding. Hopefully I'm finished with the drone section. Help me if I want to finish the 'gard' and 'npc' functions quickly or this will go on longer than Arcade does.